imported interface prefab and xr prefab from my vr room
imported interface prefab and xr rig from my room. currently disabled in the scene so that it still works.
next steps are to adapt the interface to the chemistry room adding setting to change scenes / settings. need to assign vr controller buttons to the xr rig to be able to pull up the interface.